[FEATURE] fix the ratings funcionalities
