+++ locfilm/locations/models/__init__.py
@@ -2,4 +2,5 @@ from .locations import Location
 from .categories import Category
 from .groups import Group
 from .images import Image
+from .ratings import Rating
 from .additionals import AdditionalInfo
