+++ locfilm/locations/serializers/__init__.py
@@ -1,2 +1,3 @@
 from .locations import *
 from .images import *
+from .ratings import *
